Added enemy types and awareness of his buddies
Added Light, Medium and Heavy Types.

Only Heavy enemies can punch trees to get a branch. So if there are weapons lying around, heavy enemies look for non-heavy enemies in the area. If there are less non-heavy enemies than the weapons on the ground, he goes and picks it up. Else he goes to punch the tree since they can't do it.