Added some code to reiterate the target object when chosen incorrectly
There was no straightforward way to let the agents pick up the right weapon for the right goal. As in pick up a flammable weapon for a plan that includes burn action. So had to some add code before movement takes place to make sure the enemy is going towards the right target.